requirements.txt
torch==2.1.0 -f https://download.pytorch.org/whl/cu118
torchvision==0.16.0
torchaudio==2.1.0
pillow
matplotlib
segmentation-models-pytorch
numpy<2.0
ipykernel